% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{install_menu_theme}
\alias{install_menu_theme}
\title{Install rscodeio menu bar styling}
\usage{
install_menu_theme()
}
\value{
Nothing (\code{NULL} invisibly).
}
\description{
This function overwrites RStudio's default \emph{dark} \href{https://doc.qt.io/Qt-5/stylesheet-syntax.html}{Qt Style Sheet (QSS)} files with rscodeio's customized ones
in order to properly style RStudio's menu bar which cannot be altered by \href{https://rstudio.github.io/rstudio-extensions/rstudio-theme-creation.html}{RStudio's current theming API}. The original Qt Style Sheet files are backed up and can be restored by
calling \code{\link[=uninstall_menu_theme]{uninstall_menu_theme()}}.
}
\details{
Running \code{\link[=install_themes]{install_themes(style_menu_bar = TRUE)}} (which is the default) has the same effect as running this function.
\code{install_menu_theme()} can be useful to install the rscodeio menu bar styling \emph{without} the accompanying editor themes.

Administrator privileges are required to copy the QSS files to RStudio's installation directory. On Linux, it will be asked interactively for administrator
credentials by calling \href{https://en.wikipedia.org/wiki/Polkit}{Polkit}'s \href{https://www.freedesktop.org/software/polkit/docs/latest/pkexec.1.html}{\code{pkexec}}.
Windows users must run RStudio as administrator On macOS and RStudio Server, menu bar styling is not necessary and running this function has no effect.

Note that this function has to be executed again after \emph{every} update or reinstallation of RStudio. This is because the custom QSS files provided by rscodeio
get overwritten during RStudio's installation process.
}
