% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{install_theme}
\alias{install_theme}
\title{Install the rscodeio theme}
\usage{
install_theme()
}
\value{
nothing.
}
\description{
You'll need RStudio at least 1.2.x and if you're on Windows, and your RStudio
is installed to program files, you'll need to be running RStudio as
Administrator (Just required for install).
}
