% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{uninstall_menu_theme}
\alias{uninstall_menu_theme}
\title{Uninstall rscodeio menu bar styling}
\usage{
uninstall_menu_theme()
}
\value{
Nothing (\code{NULL} invisibly).
}
\description{
This function restores RStudio's default \emph{dark} Qt \href{https://doc.qt.io/Qt-5/stylesheet-syntax.html}{Qt Style Sheet (QSS)} files.
}
\details{
Running \code{\link[=uninstall_themes]{uninstall_themes(restore_menu_bar = TRUE)}} (which is the default) has the same effect as running this function.
\code{uninstall_menu_theme()} can be useful to remove the rscodeio menu bar styling \emph{without} removing the accompanying editor themes.

Administrator privileges are required to restore RStudio's default QSS files. On Linux, it will be asked interactively for administrator credentials by
calling \href{https://en.wikipedia.org/wiki/Polkit}{Polkit}'s \href{https://www.freedesktop.org/software/polkit/docs/latest/pkexec.1.html}{\code{pkexec}}. Windows users
must run RStudio as administrator. On macOS and RStudio Server, menu bar styling is not necessary and running this function has no effect.
}
