% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{uninstall_themes}
\alias{uninstall_themes}
\title{Uninstall the rscodeio editor themes and menu bar styling}
\usage{
uninstall_themes(restore_menu_bar = TRUE)
}
\arguments{
\item{restore_menu_bar}{Whether to restore RStudio's default menu bar styling. Irrelevant if rscodeio menu bar styling is not installed (see
\code{\link[=install_menu_theme]{install_menu_theme()}} for details about menu bar styling).}
}
\value{
Nothing (\code{NULL} invisibly).
}
\description{
To restore RStudio's default \emph{dark} Qt theme (\code{.qss}) files (i.e. setting \code{restore_menu_bar = TRUE}), administrator privileges are required. On Linux, it
will be asked interactively for administrator credentials by calling \href{https://en.wikipedia.org/wiki/Polkit}{Polkit}'s
\href{https://www.freedesktop.org/software/polkit/docs/latest/pkexec.1.html}{\code{pkexec}}. Windows users must run RStudio as administrator in order to uninstall
rscodeio's menu bar styling. On macOS and RStudio Server, menu bar styling is not necessary and thus its menu bar restoration will be skipped.
}
