% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{install_theme}
\alias{install_theme}
\title{Install the rscodeio theme}
\usage{
install_theme(menus = TRUE)
}
\arguments{
\item{menus}{if FALSE do not install the RStudio menu theme qss files.}
}
\value{
nothing.
}
\description{
You'll need RStudio at least 1.2.x and if your RStudio
is installed to a default location on Windows or Linux,
you'll need to be running RStudio as
Administrator to install the menu theme files (Just required for install).
}
\details{
You can elect not to install the menu theme files with `menus = FALSE`.
}
